fix datatype bug + add example for custom query

+++ src/main/java/prm/projectbase/dto/request/UserCreateRequest.java
@@ -27,5 +27,5 @@ public class UserCreateRequest {
 
     String fullName;
 
-    Long roleId;
+    Integer roleId;
 }
